pleer_scraper/requirements.txt
grpcio-tools~=1.30
requests>=2.28.1
beautifulsoup4>=4.11.1
MarkupSafe==2.0.1
